Extract from PostgreRepository into new SubscriptionRepository
