+++ src/Altinn.Platform/Altinn.Platform.Events/Tests/TestingServices/SubscriptionServiceTest.cs
@@ -16,7 +16,7 @@ namespace Tests.TestingServices
         [Fact]
         public async Task GetOrgSubscriptions_Three_Match()
         {
-            SubscriptionService subscriptionService = new SubscriptionService(new PostgresRepositoryMock());
+            SubscriptionService subscriptionService = new SubscriptionService(new SubscriptionRepositoryMock());
             List<Subscription> result = await subscriptionService.GetOrgSubscriptions(
                 "https://ttd.apps.altinn.no/ttd/endring-av-navn-v2/",
                 "/party/1337",
@@ -27,7 +27,7 @@ namespace Tests.TestingServices
         [Fact]
         public async Task GetOrgSubscriptions_Zero_Match()
         {
-            SubscriptionService subscriptionService = new SubscriptionService(new PostgresRepositoryMock());
+            SubscriptionService subscriptionService = new SubscriptionService(new SubscriptionRepositoryMock());
             List<Subscription> result = await subscriptionService.GetOrgSubscriptions(
                 "https://ttd.apps.altinn.no/ttd/endring-av-navn-v1/",
                 "/party/1337",
@@ -38,7 +38,7 @@ namespace Tests.TestingServices
         [Fact]
         public async Task GetSubscriptions_One_Match()
         {
-            SubscriptionService subscriptionService = new SubscriptionService(new PostgresRepositoryMock());
+            SubscriptionService subscriptionService = new SubscriptionService(new SubscriptionRepositoryMock());
             List<Subscription> result = await subscriptionService.GetSubscriptions(
                 "https://ttd.apps.altinn.no/ttd/new-app/",
                 "/party/1337",
@@ -49,7 +49,7 @@ namespace Tests.TestingServices
         [Fact]
         public async Task GetSubscriptions_Zero_Match()
         {
-            SubscriptionService subscriptionService = new SubscriptionService(new PostgresRepositoryMock());
+            SubscriptionService subscriptionService = new SubscriptionService(new SubscriptionRepositoryMock());
             List<Subscription> result = await subscriptionService.GetSubscriptions(
                 "https://ttd.apps.altinn.no/ttd/endring-av-navn-v1/",
                 "/party/1337",
